chatbot/requirements.txt
langchain
langchain_community
langchain-openai
langchain_pinecone
langchain[docarray]
docarray
pydantic==1.10.8
pytubefix 
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff
pypdf
git+https://github.com/openai/whisper.git
PyPDF2
streamlit
pandas
faiss-cpu
tiktoken